Removed bootstrap with password, isPublished flag, IdentityKeyPair, Added rotatePrivateKey, restorePrivateKey
