--- Source/Authentication/AuthManager+SignIn.swift
@@ -1,65 +1,0 @@
-//
-// Copyright (C) 2015-2018 Virgil Security Inc.
-//
-// All rights reserved.
-//
-// Redistribution and use in source and binary forms, with or without
-// modification, are permitted provided that the following conditions are
-// met:
-//
-//     (1) Redistributions of source code must retain the above copyright
-//     notice, this list of conditions and the following disclaimer.
-//
-//     (2) Redistributions in binary form must reproduce the above copyright
-//     notice, this list of conditions and the following disclaimer in
-//     the documentation and/or other materials provided with the
-//     distribution.
-//
-//     (3) Neither the name of the copyright holder nor the names of its
-//     contributors may be used to endorse or promote products derived from
-//     this software without specific prior written permission.
-//
-// THIS SOFTWARE IS PROVIDED BY THE AUTHOR ''AS IS'' AND ANY EXPRESS OR
-// IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
-// WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-// DISCLAIMED. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR ANY DIRECT,
-// INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
-// (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
-// SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION)
-// HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT,
-// STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING
-// IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
-// POSSIBILITY OF SUCH DAMAGE.
-//
-// Lead Maintainer: Virgil Security Inc. <support@virgilsecurity.com>
-//
-
-import Foundation
-import VirgilCryptoApiImpl
-
-extension AuthManager {
-    internal func signInFromNewDevice(password: String, completion: @escaping (Error?) -> ()) {
-        self.cloudKeyManager.retrieve(usingPassword: password) { entry, error in
-            guard let entry = entry, error == nil else {
-                completion(error)
-                return
-            }
-            do {
-                try self.localKeyManager.store(data: entry.data, isPublished: true)
-                completion(nil)
-            } catch {
-                completion(error)
-            }
-        }
-    }
-
-    internal func signInFromKnownDevice(identityKeyPair: IdentityKeyPair, completion: @escaping (Error?) -> ()) {
-        guard !identityKeyPair.isPublished else {
-            completion(nil)
-            return
-        }
-        let keyPair = VirgilKeyPair(privateKey: identityKeyPair.privateKey, publicKey: identityKeyPair.publicKey)
-
-        self.publishCardThenUpdateLocal(keyPair: keyPair, completion: completion)
-    }
-}
